requirements.txt
# Google Agent Development Kit (ADK)
google-adk[database]==0.3.0

# Google API client libraries (compatible with google-adk)
google-api-python-client>=2.157.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-auth>=2.25.2

# Additional dependencies for document processing
python-docx==0.8.11
PyPDF2==3.0.1
markdown==3.5.1

# Web framework for backend API (compatible with google-adk)
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Environment management
python-dotenv==1.1.0

# Data processing
pandas==2.1.4
requests==2.31.0 

# Additional ADK dependencies
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
deprecated==1.2.18 